qt, refactor: Move BitcoinCore class into its own module
This change makes BitcoinCore class re-usable by an alternative GUI,
e.g., QML-based one.
